feat: auto-trigger login when not authenticated
Instead of throwing an error telling users to run `insforge login`,
commands now seamlessly start the OAuth login flow inline and continue
after authentication. Extracts reusable performOAuthLogin() into auth.ts
and makes requireAuth() async with auto-login fallback.
