+++ src/commands/db/import.ts
@@ -14,7 +14,7 @@ export function registerDbImportCommand(dbCmd: Command): void {
     .action(async (file: string, opts, cmd) => {
       const { json } = getRootOpts(cmd);
       try {
-        requireAuth();
+        await requireAuth();
         const config = getProjectConfig();
         if (!config) throw new ProjectNotLinkedError();
