+++ src/commands/db/policies.ts
@@ -26,7 +26,7 @@ export function registerDbPoliciesCommand(dbCmd: Command): void {
     .action(async (_opts, cmd) => {
       const { json } = getRootOpts(cmd);
       try {
-        requireAuth();
+        await requireAuth();
 
         const res = await ossFetch('/api/database/policies');
         const raw = await res.json() as unknown;
