+++ src/commands/db/tables.ts
@@ -11,7 +11,7 @@ export function registerDbTablesCommand(dbCmd: Command): void {
     .action(async (_opts, cmd) => {
       const { json } = getRootOpts(cmd);
       try {
-        requireAuth();
+        await requireAuth();
 
         const res = await ossFetch('/api/database/tables');
         const tables = await res.json() as string[];
