+++ src/commands/db/triggers.ts
@@ -26,7 +26,7 @@ export function registerDbTriggersCommand(dbCmd: Command): void {
     .action(async (_opts, cmd) => {
       const { json } = getRootOpts(cmd);
       try {
-        requireAuth();
+        await requireAuth();
 
         const res = await ossFetch('/api/database/triggers');
         const raw = await res.json() as unknown;
