+++ src/commands/functions/invoke.ts
@@ -13,7 +13,7 @@ export function registerFunctionsInvokeCommand(functionsCmd: Command): void {
     .action(async (slug: string, opts, cmd) => {
       const { json } = getRootOpts(cmd);
       try {
-        requireAuth();
+        await requireAuth();
 
         const config = getProjectConfig();
         if (!config) throw new ProjectNotLinkedError();
