+++ src/commands/projects/link.ts
@@ -26,7 +26,7 @@ export function registerProjectLinkCommand(program: Command): void {
     .action(async (opts, cmd) => {
       const { json, apiUrl } = getRootOpts(cmd);
       try {
-        requireAuth();
+        await requireAuth(apiUrl);
 
         let orgId = opts.orgId;
         let projectId = opts.projectId;
