+++ src/commands/schedules/create.ts
@@ -17,7 +17,7 @@ export function registerSchedulesCreateCommand(schedulesCmd: Command): void {
     .action(async (opts, cmd) => {
       const { json } = getRootOpts(cmd);
       try {
-        requireAuth();
+        await requireAuth();
 
         const body: Record<string, unknown> = {
           name: opts.name,
