+++ src/commands/storage/buckets.ts
@@ -11,7 +11,7 @@ export function registerStorageBucketsCommand(storageCmd: Command): void {
     .action(async (_opts, cmd) => {
       const { json } = getRootOpts(cmd);
       try {
-        requireAuth();
+        await requireAuth();
 
         const res = await ossFetch('/api/storage/buckets');
         const raw = await res.json() as unknown;
